# swagger/swagger.yml
swagger: "2.0"
info:
  description: "Servicios que expone el servidor para el consumo mediante HTTP a través de clientes Android y CIAA."
  version: "1.0.0"
  title: "Micrófono Remoto"
  contact:
    email: "augustobonifacio95@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "dispositivo-ciaa"
  description: "Servicios a consumir por la CIAA"
- name: "dispositivo-cliente"
  description: "Servicios a consumir por los dispositivos de escucha"
schemes:
- "http"
paths:
  /dispositivos:
    post:
      tags:
      - "dispositivo-ciaa"
      summary: "Registra un micrófono"
      operationId: "addMic"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Datos del micrófono a registrar"
        required: true
        schema:
          $ref: "#/definitions/Dispositivo"
      responses:
        200:
          description: "Devuele el puerto que el servidor abre para recibir la información del micrófono"
          schema:
            $ref: "#/definitions/Puerto"
        500:
          description: "Error. En el Header 'Error-Message' se encuentra la causa"
    put:
      tags:
      - "dispositivo-cliente"
      summary: "Solicita escuchar o dejar de escuchar a un determinado micrófono"
      operationId: "listenMic"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Solicitud de escucha"
        required: true
        schema:
          $ref: "#/definitions/Solicitud"
      responses:
        200:
          description: "Si la solicitud resulta OK"
        500:
          description: "Error. En el Header 'Error-Message' se encuentra la causa"
    get:
      tags:
      - "dispositivo-cliente"
      summary: "Lista los micrófonos registrados"
      operationId: "getMics"
      produces:
      - "application/json"
      parameters:
      - name: "all"
        in: "query"
        description: "Si es true devuelve todos los dispositivos registrados, caso contrario los que esten transmitiendo"
        required: true
        type: "boolean"
      responses:
        200:
          description: "Devuelve todos los dispositivos registrados o aquellos que estan transmitiendo actualmente"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Dispositivo"
        500:
          description: "Error. En el Header 'Error-Message' se encuentra la causa"
  /dispositivos/{mac}:
    get:
      tags:
      - "dispositivo-ciaa"
      summary: "Devuelve el detalle de un dispositivo con sus archivos .wav históricos"
      operationId: "getMic"
      produces:
      - "application/json"
      parameters:
      - name: "mac"
        in: "path"
        description: "MAC del micrófono"
        required: true
        type: "string"
      responses:
        200:
          description: "Devuelve todos los datos del dispositivo junto con una lista rutas a los archivos .wav históricos"
          schema:
            $ref: "#/definitions/Dispositivo"
        500:
          description: "Error. En el Header 'Error-Message' se encuentra la causa"
    delete:
      tags:
      - "dispositivo-ciaa"
      summary: "Da de baja un micrófono"
      operationId: "deleteMic"
      produces:
      - "application/json"
      parameters:
      - name: "mac"
        in: "path"
        description: "MAC del micrófono"
        required: true
        type: "string"
      responses:
        200:
          description: "Se dió de baja el dispositivo"
        500:
          description: "Error. En el Header 'Error-Message' se encuentra la causa"
definitions:
  Puerto:
    type: "number"
    format: "unit8"
    example: 8081
  Dispositivo:
    type: "object"
    required:
      - "nombre"
      - "mac"
      - "sampleRate"
      - "sampleSize"
    properties:
      mac:
        type: "string"
        example: "ad:af:ac:65:b3:c9"
        description: "Ip del dispostivo"
      nombre:
        type: "string"
        example: "CIAA"
        description: "Nombre del dispositivo"
      subscribers:
        type: "array"
        description: "Lista de clientes escuchando a este dispositivo"
        items:
          type: "string"
          description: "Ip de quien escucha"
          example: "192.168.0.200"
      sampleRate:
        type: "number"
        format: "uint8"
        example: 44000
        description: "Frecuencia de muestreo del audio transmitido"
      sampleSize:
        type: "number"
        format: "uint8"
        example: 8
        description: "Tamaño de la muestra del audio transmitido: (8 | 16) bits"
      stereo:
        type: "boolean"
        example: true
        description: "Si transmite en stereo (por defecto false)"
      puerto:
        type: "number"
        format: "uint8"
        description: "Puerto del servidor al cuál transmite el dispositivo"
      files:
        type: "array"
        description: "Archivos .wav históricos"
        items:
          $ref: "#/definitions/File"
  File:
    type: "object"
    properties:
      name:
        type: "string"
        example: "22-11-17 1434"
        description: "Nombre del archivo con fecha y hora de creación"
      path:
        type: "string"
        description: "Ruta del archivo"
        example: "/records/ab:ac:ad:19:29:39/22-11-17T1434.wav"
  Solicitud:
    type: "object"
    required:
      - "mac"
      - "ip"
    properties:
      ip:
        type: "string"
        example: "192.168.0.10"
        description: "Ip del dispostivo solicitante"
      mac:
        type: "string"
        description: "MAC del micrófono"
        example: "ab:ac:ad:19:29:39"